+++ packages/components/organisms/f-takeawaypay-activation/CHANGELOG.md
@@ -3,13 +3,19 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.3.1
+------------------------------
+*October 08, 2021*
+
+### Changed
+- Component tests
+- Accessibility tests
 
 v2.2.1
 ------------------------------
 *October 06, 2021*
 
 Re-publish to clean up the old styles from the compiled package.
-
 
 v2.2.0
 ------------------------------